Always run stack-allocation pass regardless of the optimization level
Previously, in non-optimized mode `EVMVRegToMem` pass is ran.
But this optimization seems to quite buggy, it immediately fails on
test_stdlib.cpp, when passing `-O0` to the command line.
Speaking about this legacy, it is hard to find out any reason why we
shouldn't run stack allocation in an unoptimized mode.