Support new memory instructions and native alignment
All frontend fixes related to 256-bit alignment were removed.

The main components affected by the new aligment:
- Lowering: several nodes require custom lowering due to operands with unsuitable
  types(not i256).
- Stack allocation: it still uses 256-bits slots for spill/fills, but memory for
  the local data could not be aligned in 256-bits. Thus, it should be taken into
  account and properly handled by EVMStackAllocation and EVMConvertRegToStack.
- MC and Linker: now we can't treat all integers as 256-bits, therefore the MC
  emitter produces object files with more detailed description of data. Linker
  in turn supports this new description and support init data, which is not
  256-bit aligned now.

Also added ability to use `evmone` as a VM for testing. But it has a drawback:
it doesn't support precompiled contracts, so test for __evm_builtin_modpow is
disabled for a while.

Also, due to proper data size and alignment, `memset` function is working
correctly now. So it's test was enabled.

Signed-off-by: Mikhail Sherstennikov <mike.sherstennikov@gmail.com>